Fixed overflow issue
Introduced BigDecimal to avoid overflow

+++ src/main/java/calculator/operator/IOperator.java
@@ -1,8 +1,9 @@
 package calculator.operator;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface IOperator {
 
-    double apply(Map<String, Double> contextValues) throws Exception;
+    BigDecimal apply(Map<String, BigDecimal> contextValues) throws Exception;
 }
